refactored app.py into sub .py files in lib directory
